feat: Create base form
