--- app/prodcon/programs.ts
@@ -1,514 +1,0 @@
-export const programs = 
-[
-    {
-      "code": "ACC",
-      "name": "ACC",
-      "description": null,
-      "descriptionFormal": "Accounting & Finance - School of",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "AMATH",
-      "name": "AMATH",
-      "description": null,
-      "descriptionFormal": "Applied Mathematics",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "ANTH",
-      "name": "ANTH",
-      "description": null,
-      "descriptionFormal": "Anthropology",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "ARCH",
-      "name": "ARCH",
-      "description": null,
-      "descriptionFormal": "Architecture - School of",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "ART",
-      "name": "ART",
-      "description": null,
-      "descriptionFormal": "Arts",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "ARTSDEAN",
-      "name": "ARTSDEAN",
-      "description": null,
-      "descriptionFormal": "Dean of Arts",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "BIOL",
-      "name": "BIOL",
-      "description": null,
-      "descriptionFormal": "Biology",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "CBET",
-      "name": "CBET",
-      "description": null,
-      "descriptionFormal": "Conrad Business, Entrepreneurship & Technology Ctr",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "CECS",
-      "name": "CECA",
-      "description": null,
-      "descriptionFormal": "Co-operative Education & Career Action",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "CGC",
-      "name": "CGC",
-      "description": null,
-      "descriptionFormal": "Conrad Grebel University College",
-      "associatedCampusCode": "CGC"
-    },
-    {
-      "code": "CHE",
-      "name": "CH E",
-      "description": null,
-      "descriptionFormal": "Chemical Engineering",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "CHEM",
-      "name": "CHEM",
-      "description": null,
-      "descriptionFormal": "Chemistry",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "CIVE",
-      "name": "CIVE",
-      "description": null,
-      "descriptionFormal": "Civil and Environmental Engineering",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "CLASSICS",
-      "name": "CLASSICS",
-      "description": null,
-      "descriptionFormal": "Classical Studies",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "CO",
-      "name": "C&O",
-      "description": null,
-      "descriptionFormal": "Combinatorics & Optimization",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "CS",
-      "name": "CS",
-      "description": null,
-      "descriptionFormal": "Computer Science - David R. Cheriton School of",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "EARTH",
-      "name": "EARTHENV",
-      "description": null,
-      "descriptionFormal": "Earth and Environmental Sciences",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "ECE",
-      "name": "E&CE",
-      "description": null,
-      "descriptionFormal": "Electrical & Computer Engineering",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "ECON",
-      "name": "ECON",
-      "description": null,
-      "descriptionFormal": "Economics",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "ENG",
-      "name": "ENG",
-      "description": null,
-      "descriptionFormal": "Engineering",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "ENGDEAN",
-      "name": "ENGDEAN",
-      "description": null,
-      "descriptionFormal": "Dean of Engineering",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "ENGL",
-      "name": "ENGL",
-      "description": null,
-      "descriptionFormal": "English Language & Literature",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "ENV",
-      "name": "ENV",
-      "description": null,
-      "descriptionFormal": "Environment",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "ENVDEAN",
-      "name": "ENVDEAN",
-      "description": null,
-      "descriptionFormal": "Dean of Environment",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "ERS",
-      "name": "ERS",
-      "description": null,
-      "descriptionFormal": "Environment, Resources & Sustainability, School of",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "FINE",
-      "name": "FINE",
-      "description": null,
-      "descriptionFormal": "Fine Arts",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "FR",
-      "name": "FR",
-      "description": null,
-      "descriptionFormal": "French Studies",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "GEOG",
-      "name": "GEOG",
-      "description": null,
-      "descriptionFormal": "Geography & Environmental Management",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "GERSLAV",
-      "name": "GER&SLAV",
-      "description": null,
-      "descriptionFormal": "Germanic & Slavic Studies",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "GRAD",
-      "name": "GRAD",
-      "description": null,
-      "descriptionFormal": "Graduate Studies",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "GRADDEAN",
-      "name": "Grad Dean",
-      "description": null,
-      "descriptionFormal": "Graduate Dean",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "HIST",
-      "name": "HIST",
-      "description": null,
-      "descriptionFormal": "History",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "ICR",
-      "name": "ICR",
-      "description": null,
-      "descriptionFormal": "Institute for Computer Research",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "IS",
-      "name": "IS",
-      "description": null,
-      "descriptionFormal": "Independent Studies",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "KI",
-      "name": "KI",
-      "description": null,
-      "descriptionFormal": "Knowledge Integration",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "MAT",
-      "name": "MAT",
-      "description": null,
-      "descriptionFormal": "Math",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "MATHDEAN",
-      "name": "MATHDEAN",
-      "description": null,
-      "descriptionFormal": "Dean of Mathematics",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "ME",
-      "name": "ME",
-      "description": null,
-      "descriptionFormal": "Mechanical and Mechatronics Engineering",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "OPTOM",
-      "name": "OPTOM",
-      "description": null,
-      "descriptionFormal": "Optometry & Vision Science - School of",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "PHARM",
-      "name": "PHARM",
-      "description": null,
-      "descriptionFormal": "Pharmacy - School of",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "PHIL",
-      "name": "PHIL",
-      "description": null,
-      "descriptionFormal": "Philosophy",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "PHYS",
-      "name": "PHYSASTR",
-      "description": null,
-      "descriptionFormal": "Physics & Astronomy",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "PLAN",
-      "name": "PLAN",
-      "description": null,
-      "descriptionFormal": "Planning - School of",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "PMATH",
-      "name": "PMATH",
-      "description": null,
-      "descriptionFormal": "Pure Mathematics",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "PSCI",
-      "name": "PSCI",
-      "description": null,
-      "descriptionFormal": "Political Science",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "PSYCH",
-      "name": "PSYCH",
-      "description": null,
-      "descriptionFormal": "Psychology",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "REC",
-      "name": "REC",
-      "description": null,
-      "descriptionFormal": "Recreation & Leisure Studies",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "REN",
-      "name": "REN",
-      "description": null,
-      "descriptionFormal": "Renison University College",
-      "associatedCampusCode": "REN"
-    },
-    {
-      "code": "RS",
-      "name": "RS",
-      "description": null,
-      "descriptionFormal": "Religious Studies",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "SCI",
-      "name": "SCI",
-      "description": null,
-      "descriptionFormal": "Science",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "SCIDEAN",
-      "name": "SCIDEAN",
-      "description": null,
-      "descriptionFormal": "Dean of Science Office",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "SEED",
-      "name": "SEED",
-      "description": null,
-      "descriptionFormal": "Environment, Enterprise & Development - School of",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "SOC",
-      "name": "SOC",
-      "description": null,
-      "descriptionFormal": "Sociology and Legal Studies",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "SPAN",
-      "name": "SPAN",
-      "description": null,
-      "descriptionFormal": "Spanish & Latin American Studies",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "STATACTSC",
-      "name": "STAT&ACTSC",
-      "description": null,
-      "descriptionFormal": "Statistics & Actuarial Science",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "STJ",
-      "name": "STJ",
-      "description": null,
-      "descriptionFormal": "St Jerome's University",
-      "associatedCampusCode": "STJ"
-    },
-    {
-      "code": "STV",
-      "name": "STV",
-      "description": null,
-      "descriptionFormal": "Society, Technology & Values Centre",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "SYDE",
-      "name": "SY DE",
-      "description": null,
-      "descriptionFormal": "Systems Design Engineering",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "THC",
-      "name": "THC",
-      "description": null,
-      "descriptionFormal": "Theatre Centre",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "UW",
-      "name": "UW",
-      "description": null,
-      "descriptionFormal": "University of Waterloo",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "VPA",
-      "name": "VPA",
-      "description": null,
-      "descriptionFormal": "Interdisciplinary Studies",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "WS",
-      "name": "W S",
-      "description": null,
-      "descriptionFormal": "Women's Studies",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "SIDB",
-      "name": "SIDB",
-      "description": null,
-      "descriptionFormal": "Stratford School of Interaction Design & Business",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "CBET",
-      "name": "CBET",
-      "description": null,
-      "descriptionFormal": "Conrad School of Entrepreneurship and Business",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "DRAMA",
-      "name": "DRAMA",
-      "description": null,
-      "descriptionFormal": "Communication Arts",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "HLTHGERON",
-      "name": "PUBHLTHSCI",
-      "description": null,
-      "descriptionFormal": "Public Health Sciences - School of",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "KIN",
-      "name": "KINHLTHSCI",
-      "description": null,
-      "descriptionFormal": "Kinesiology and Health Sciences",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "AHS",
-      "name": "AHS",
-      "description": null,
-      "descriptionFormal": "Health",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "AHSDEAN",
-      "name": "AHSDEAN",
-      "description": null,
-      "descriptionFormal": "Dean of Health",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "AHS",
-      "name": "AHS",
-      "description": null,
-      "descriptionFormal": "Health",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "AHSDEAN",
-      "name": "AHSDEAN",
-      "description": null,
-      "descriptionFormal": "Dean of Health",
-      "associatedCampusCode": "UW"
-    },
-    {
-      "code": "STP",
-      "name": "STP",
-      "description": null,
-      "descriptionFormal": "United College",
-      "associatedCampusCode": "STP"
-    },
-    {
-      "code": "MSCI",
-      "name": "M SCI",
-      "description": null,
-      "descriptionFormal": "Management Sciences and Engineering",
-      "associatedCampusCode": "UW"
-    }
-]
